+++ tests/test_transport.py
@@ -3,12 +3,41 @@
 # MIT License (see LICENSE or https://opensource.org/licenses/MIT)
 
 import pytest
+
 import re
+import socket
+import struct
+import threading
+import time
 
 from smbprotocol.transport import (
+    _TimeoutError,
     DirectTCPPacket,
     Tcp,
 )
+
+
+@pytest.fixture()
+def server_tcp(request):
+    server_func_name = 'server_' + request.node.name
+    server_func = globals().get(server_func_name)
+    if not server_func:
+        raise Exception('Test must have defined %s to run the server thread' % server_func_name)
+
+    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    try:
+        sock.bind(('127.0.0.1', 0))
+        sock.listen(1)
+
+        server_thread = threading.Thread(target=server_func, args=(sock,))
+        server_thread.start()
+        try:
+            yield Tcp(*sock.getsockname())
+
+        finally:
+            server_thread.join()
+    finally:
+        sock.close()
 
 
 class TestDirectTcpPacket(object):
@@ -41,8 +70,8 @@ class TestDirectTcpPacket(object):
 class TestTcp(object):
 
     def test_normal_fail_message_too_big(self):
-        tcp = Tcp("0.0.0.0", 0, None)
-        tcp._connected = True
+        tcp = Tcp("0.0.0.0", 0)
+        tcp.connected = True
         with pytest.raises(ValueError) as exc:
             tcp.send(b"\x00" * 16777216)
         assert str(exc.value) == "Data to be sent over Direct TCP size " \
@@ -50,6 +79,98 @@ class TestTcp(object):
                                  "16777215"
 
     def test_invalid_host(self):
-        tcp = Tcp("fake-host", 445, None)
+        tcp = Tcp("fake-host", 445)
         with pytest.raises(ValueError, match=re.escape("Failed to connect to 'fake-host:445': ")):
-            tcp.send(b"")
+            tcp.connect()
+
+
+def test_small_recv(server_tcp):
+    server_tcp.connect()
+    server_tcp.send(b'\x00')
+
+    actual = server_tcp.recv(10)
+
+    server_tcp.send(b'\x00')
+
+    assert actual == b"\x01\x02\x03\x04"
+
+
+def server_test_small_recv(server):
+    # I'm not sure how else to test this but it seems like the small sleeps is enough for the client recv() to read it
+    # in the chunks we send.
+    sock = server.accept()[0]
+    try:
+        sock.recv(5)
+
+        b_len = struct.pack(">I", 4)
+
+        sock.send(b_len[:2])
+        time.sleep(0.1)
+        sock.send(b_len[2:])
+
+        sock.send(b'\x01\x02')
+        time.sleep(0.1)
+        sock.send(b'\x03\x04')
+
+        sock.recv(5)
+
+    finally:
+        sock.shutdown(socket.SHUT_RDWR)
+        sock.close()
+
+
+def test_recv_timeout(server_tcp):
+    server_tcp.connect()
+
+    with pytest.raises(_TimeoutError):
+        server_tcp.recv(1)
+
+    server_tcp.send(b'\x00')
+
+
+def server_test_recv_timeout(server):
+    sock = server.accept()[0]
+    try:
+        sock.recv(5)
+
+    finally:
+        sock.shutdown(socket.SHUT_RDWR)
+        sock.close()
+
+
+def test_recv_closed(server_tcp):
+    server_tcp.connect()
+    actual = server_tcp.recv(10)
+    assert actual == b''
+    assert server_tcp.connected is False
+
+
+def server_test_recv_closed(server):
+    sock = server.accept()[0]
+    try:
+        time.sleep(1)
+
+    finally:
+        sock.shutdown(socket.SHUT_RDWR)
+        sock.close()
+
+
+def test_recv_closed_client(server_tcp):
+    server_tcp.connect()
+    recv_res = []
+
+    def recv():
+        recv_res.append(server_tcp.recv(5))
+    client_recv_t = threading.Thread(target=recv)
+    client_recv_t.start()
+
+    server_tcp.close()
+    client_recv_t.join()
+    assert recv_res == [b'']
+
+
+def server_test_recv_closed_client(server):
+    sock = server.accept()[0]
+    sock.recv(1)
+    sock.shutdown(socket.SHUT_RDWR)
+    sock.close()